# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  # TO DO: CHANGE THE DESCRIPTION
  description: |-
    This OpenAPI document describes the application of a software which allows people to stay in touch with other users of the system.
    With the inspiration of `WhatsApp`, this app goals is to create an
    open-source one with less security in it
  version: "1"
  contact:
    name: Diego Antonio Guzman Aguirre
    email: guzmanaguirre.2006301@studenti.uniroma1.it
    url: https://github.com/guizzyy
servers:
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /users/{uID}/username:
    parameters:
      - $ref: '#/components/parameters/UserID'

    put:
      tags: ["user profile"]
      summary: set a new username
      description: |
        Allows the user to set a new username for his account
      operationId: setMyUserName
      requestBody:
        description: the new username
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Username updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /users/{uID}/photo:
    parameters:
      - $ref: '#/components/parameters/UserID'

    put:
      tags: ["user profile"]
      summary: set a new photo
      description: |
        Allows the user to set a new photo for his account
      operationId: setMyPhoto
      requestBody:
        description: the new photo
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: new photo updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /conversations/{convID}/manage/name:
    parameters:
      - $ref: '#/components/parameters/ConversationID'

    put:
      tags: ["groups"]
      summary: set a new group name
      description: |
        Change the current group name with another one
      operationId: setGroupName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: group name changed correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /conversations/{convID}/manage/photo:
    parameters:
      - $ref: '#/components/parameters/ConversationID'

    put:
      tags: ["groups"]
      summary: set a new group photo
      description: |
        Update a new photo for the group
      operationId: setGroupPhoto
      requestBody:
        description: the new group photo
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: new group photo updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /conversations/{convID}/messages:
    parameters:
      - $ref: '#/components/parameters/ConversationID'

    post:
      tags: ["chats"]
      summary: send a new message
      description: |
        Send a new message in a conversation with a user or a group
      operationId: sendMessage
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Message'
        required: true
      responses:
        '200':
          description: message correctly sent
          content:
            multipart/form-data:
              schema:
                $ref: '#/components/schemas/Message'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      description: authentication code mode
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
    Unauthorized:
      description: The access token is missing or it's expired
    NotFound:
      description: Resource not found in the database
  parameters:
    UserID:
      schema:
        type: integer
        example: 1
      name: uID
      in: path
      required: true
      description: ID of the user
    ConversationID:
      schema:
        type: integer
        example: 2
      name: convID
      in: path
      required: true
      description: ID of the conversation
...
  schemas:
    User:
      description: User schema
      type: object
      properties:
        id:
          type: integer
          example: 10
          readOnly: true
          description: user ID
        username:
          type: string
          example: Diego
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: username
        photo:
          $ref: '#/components/schemas/Photo'
    Group:
      description: group schema
      type: object
      properties:
        id:
          type: integer
          example: 11
          readOnly: true
          description: group ID
        name:
          type: string
          example: Class
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: group name
        photo:
          $ref: '#/components/schemas/Photo'
    Message:
      description: message schema
      type: object
      properties:
        id:
          type: integer
          example: 12
          readOnly: true
          description: message ID
        text:
          $ref: '#/components/schemas/Text'
        timestamp:
          type: string
          format: date-time
          minLength: 1
          maxLength: 999
          description: date and hour of the message
        sender:
          $ref: '#/components/schemas/User'
        photo:
          $ref: '#/components/schemas/Photo'
        status:
          type: string
          enum:
            - "Unreceived"
            - "Received"
            - "Read"
          description: status of the message
        reactions:
          $ref: '#/components/schemas/List-of-Reaction'
...
    Reaction:
      description: reaction schema
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        messageId:
          type: integer
          example: 1234
          description: id of the message reacted
          readOnly: true
        emoji:
          type: string
          example: "🤣"
          minLength: 1
          maxLength: 4
          pattern: "^[\\u1F600-\\u1F64F\\u1F300-\\u1F5FF\\u1F680-\\u1F6FF\\u1F700-\\u1F77F\\u1F780-\\u1F7FF\\u1F800-\\u1F8FF\\u1F900-\\u1F9FF\\u1FA00-\\u1FA6F\\u1FA70-\\u1FAFF\\u2600-\\u26FF\\u2700-\\u27BF\\u2300-\\u23FF\\u2B50\\u23F0\\u231A\\u25AA-\\u25FE\\u2B06\\u2194-\\u2199\\u21A9-\\u21AA\\u2753-\\u2755\\u274C\\u274E]$"
          description: single reacted emoji
...
    List-of-Reaction:
      description: List of reaction schema
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'
          minItems: 0
          maxItems: 999
          description: reactions found
...
    Photo:
      description: Photo schema
      type: object
      properties:
        file:
          type: string
          format: binary
          minLength: 1
          maxLength: 999
          description: file of the photo
    Text:
      description: message text
      type: object
      properties:
        text:
          type: string
          example: This is a message!
          pattern: '^.*?$'
          minLength: 1
          maxLength: 255
          description: the text of the message
    Notification:
      description: notification sent to the user
      type: object
      properties:
        report:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 250
          example: "Resource updated successfully"
          description: notify user of the operation result